finish

+++ database/migrations/2025_06_29_060413_add_harga_member_to_pelanggans_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('pelanggans', function (Blueprint $table) {
-            $table->decimal('harga_member', 10, 2)->default(100000)->after('point');
+            $table->integer('harga_member')->default(100000)->after('point');
         });
     }
